requirements.txt
tensorflow
numpy
pandas
openpyxl
xlrd
streamlit
python-Levenshtein